change testing background for profile page

+++ prod_break/routes.py
@@ -1,17 +1,10 @@
-from flask import render_template, url_for, flash, redirect, request
+from flask import render_template, url_for, flash, redirect, request, abort
 from prod_break import app, bcrypt, db  # imports from __init__.py
 from prod_break.forms import RegistrationForm, LoginForm, UpdateProfileForm
 from prod_break.models import User, Task
 from flask_login import login_user, current_user, logout_user, login_required
 
 from datetime import date
-
-# tasks = [
-#     {"date": "12-21-2020", "task_name": "get groceries", "complete": False},
-#     {"date": "12-22-2020", "task_name": "walk dog", "complete": True},
-#     {"date": "12-23-2020", "task_name": "exercise", "complete": False},
-# ]
-
 
 @app.route("/")
 def home():
@@ -95,7 +88,8 @@ def add_task():
 @login_required
 def delete_task(task_id):
     task = Task.query.get(task_id)
-    # add check to make sure it is a user task
+    if not task or task.user_id != current_user.id:
+        abort(403)
     db.session.delete(task)
     db.session.commit()
     return redirect(url_for('task'))
